service.py: Fix rar files extraction
The plugin was not extracting rar files correctly. Only the first file
was being extracted. This is an __ungly__ and temporary fix for this. It
executes the unrar utility as a subprocess in a shell. Hence, for this
to work, the *unrar* command must be installed in your system. It should
only work for Linux. Extracting zip files was not tested and uses a
similar method as the original one.

Signed-off-by: Nuno Sa <noname.nuno@gmail.com>